refactor: add the custom hook useLogIn
