+++ client/src/components/Auth/Auth.jsx
@@ -13,40 +13,18 @@ import {
   AlertTitle,
   AlertDescription,
 } from "@chakra-ui/react";
-import { useForm } from "react-hook-form";
-import { users, useUserStore } from "../../store";
-import { useNavigate } from "react-router-dom";
-import { routes } from "../../config";
+
+import { useLogIn } from "../../hooks";
 
 export default function Auth() {
   const {
     handleSubmit,
-    formState: { errors, isSubmitting, isValid },
+    errors,
+    isSubmitting,
+    isValid,
     register,
-    setError,
-  } = useForm();
-
-  const logIn = useUserStore((state) => state.logIn);
-
-  const navigate = useNavigate();
-
-  const onSubmitHandler = async (data) => {
-    return new Promise((resolve) => {
-      setTimeout(() => {
-        const userFound = users.find((user) => user.email === data.email);
-        if (userFound) {
-          logIn(userFound);
-          navigate(routes.AUTH); // esto deberia redirigir a la ruta de la app, (aun no existe por lo que redirige a la ruta de atenticacion)
-        } else {
-          setError("email", {
-            type: "manual",
-            message: "correo o contraseña incorrectos",
-          });
-        }
-        resolve();
-      }, 1500);
-    });
-  };
+    onSubmitHandler,
+  } = useLogIn();
 
   return (
     <Stack
